+++ Plugins/MultiplayerSessions/Source/MultiplayerSessions.Build.cs
@@ -42,7 +42,7 @@ public class MultiplayerSessions : ModuleRules
 				"CoreUObject",
 				"Engine",
 				"Slate",
-				"SlateCore",
+				"SlateCore"
 				// ... add private dependencies that you statically link with here ...	
 			}
 			);
