+++ Plugins/MultiplayerSessions/Source/Private/Menu.cpp
@@ -1,0 +1,245 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "Menu.h"
+#include "Components/Button.h"
+#include "MultiplayerSessionsSubsystem.h"
+#include "OnlineSessionSettings.h"
+#include "OnlineSubsystem.h"
+
+
+
+void UMenu::MenuSetup(int32 NumberOfPublicConnections, FString TypeOfMatch, FString LobbyPath)
+{
+    PathToLobby = FString::Printf(TEXT("%s?listen"), *LobbyPath);
+    NumPublicConnections = NumberOfPublicConnections;
+    MatchType = TypeOfMatch;
+	AddToViewport();
+    SetVisibility(ESlateVisibility::Visible);
+    SetIsFocusable(true);
+
+    UWorld* World = GetWorld();
+    if (World)
+    {
+        APlayerController* PlayerController = World->GetFirstPlayerController();
+        if (PlayerController)
+        {
+            FInputModeUIOnly InputModeData;
+            InputModeData.SetWidgetToFocus(TakeWidget());
+            InputModeData.SetLockMouseToViewportBehavior(EMouseLockMode::DoNotLock);
+            PlayerController->SetInputMode(InputModeData);
+            PlayerController->SetShowMouseCursor(true);
+        }
+    }
+
+    UGameInstance* GameInstance = GetGameInstance();
+    if (GameInstance)
+	{
+		MultiplayerSessionsSubsystem = GameInstance->GetSubsystem<UMultiplayerSessionsSubsystem>();
+	}
+
+    if (MultiplayerSessionsSubsystem)
+    {
+        MultiplayerSessionsSubsystem->MultiplayerOnCreateSessionComplete.AddDynamic(this, &ThisClass::OnCreateSession);
+        MultiplayerSessionsSubsystem->MultiplayerOnFindSessionsComplete.AddUObject(this, &ThisClass::OnFindSessions);
+        MultiplayerSessionsSubsystem->MultiplayerOnJoinSessionComplete.AddUObject(this, &ThisClass::OnJoinSession);
+        MultiplayerSessionsSubsystem->MultiplayerOnDestroySessionComplete.AddDynamic(this, &ThisClass::OnDestroySession);
+        MultiplayerSessionsSubsystem->MultiplayerOnStartSessionComplete.AddDynamic(this, &ThisClass::OnStartSession);
+    }
+}
+
+void UMenu::HostButtonClicked()
+{
+    HostButton->SetIsEnabled(false);
+	if (GEngine)
+    {
+        GEngine->AddOnScreenDebugMessage(
+            -1,
+            15.f,
+            FColor::Yellow,
+            FString(TEXT("Host button clicked"))
+        );
+    }
+
+    if (MultiplayerSessionsSubsystem)
+    {
+        MultiplayerSessionsSubsystem->CreateSession(NumPublicConnections, MatchType);
+    }
+}
+
+void UMenu::JoinButtonClicked()
+{
+    JoinButton->SetIsEnabled(false);
+    if (GEngine)
+        {
+            GEngine->AddOnScreenDebugMessage(
+                -1,
+                15.f,
+                FColor::Red,
+                FString(TEXT("Join Button Clicked!"))
+            );
+        }
+    if (MultiplayerSessionsSubsystem)
+    {
+        MultiplayerSessionsSubsystem->FindSessions(10000);
+    }
+}
+
+bool UMenu::Initialize()
+{
+    if (!Super::Initialize())
+    {
+        return false;
+    }
+
+    if (HostButton)
+    {
+        HostButton->OnClicked.AddDynamic(this, &UMenu::HostButtonClicked);
+    }
+
+    if (JoinButton)
+    {
+        JoinButton->OnClicked.AddDynamic(this, &UMenu::JoinButtonClicked);
+    }
+
+	return true;
+}
+
+void UMenu::MenuTearDown()
+{
+	RemoveFromParent();
+    UWorld* World = GetWorld();
+    if (World)
+    {
+        APlayerController* PlayerController = World->GetFirstPlayerController();
+        if (PlayerController)
+        {
+            FInputModeGameOnly InputModeData;
+            PlayerController->SetInputMode(InputModeData);
+            PlayerController->SetShowMouseCursor(false);
+        }
+    }
+}
+
+void UMenu::NativeDestruct()
+{
+	MenuTearDown();
+
+    Super::NativeDestruct();
+}
+
+void UMenu::OnCreateSession(bool bWasSuccessful)
+{
+    if (bWasSuccessful)
+    {
+        if (GEngine)
+        {
+            GEngine->AddOnScreenDebugMessage(
+                -1,
+                15.f,
+                FColor::Yellow,
+                FString(TEXT("Session Created Successfully!"))
+            );
+        }
+        HostButton->SetIsEnabled(true);
+
+        UWorld* World = GetWorld();
+        if (World)
+        {
+            World->ServerTravel(PathToLobby);
+        }
+    }
+    else
+    {
+        if (GEngine)
+        {
+            GEngine->AddOnScreenDebugMessage(
+                -1,
+                15.f,
+                FColor::Red,
+                FString(TEXT("Session Creation Failed!"))
+            );
+        }
+    }
+}
+
+void UMenu::OnFindSessions(const TArray<FOnlineSessionSearchResult>& SessionResults, bool bWasSuccessful)
+{
+    if (GEngine)
+        {
+            GEngine->AddOnScreenDebugMessage(
+                -1,
+                15.f,
+                FColor::Green,
+                FString(TEXT("We Might've Found Sessions"))
+            );
+        }
+    if (MultiplayerSessionsSubsystem == nullptr)
+    {
+        if (GEngine)
+        {
+            GEngine->AddOnScreenDebugMessage(
+                -1,
+                15.f,
+                FColor::Green,
+                FString(TEXT("MultiplayerSessionsSubstystem doesn't exist"))
+            );
+        }
+        return;
+    }
+    for (auto Result : SessionResults)
+    {
+        
+        FString SettingsValue;
+        Result.Session.SessionSettings.Get(FName("MatchType"), SettingsValue);
+        if (SettingsValue == MatchType)
+        {
+            MultiplayerSessionsSubsystem->JoinSession(Result);
+            return;
+        }
+    }
+    if (GEngine)
+        {
+            GEngine->AddOnScreenDebugMessage(
+                -1,
+                15.f,
+                FColor::Red,
+                FString(TEXT("We Might've Not Actually"))
+            );
+        }
+    JoinButton->SetIsEnabled(true);
+}
+
+void UMenu::OnJoinSession(EOnJoinSessionCompleteResult::Type Result)
+{
+    IOnlineSubsystem* Subsystem = IOnlineSubsystem::Get();
+    if (Subsystem)
+    {
+        IOnlineSessionPtr SessionInterface = Subsystem->GetSessionInterface();
+        if (SessionInterface.IsValid())
+        {
+            FString Address;
+            SessionInterface->GetResolvedConnectString(NAME_GameSession, Address);
+
+            APlayerController* PlayerController = GetGameInstance()->GetFirstLocalPlayerController();
+            if (PlayerController)
+            {
+                PlayerController->ClientTravel(Address, ETravelType::TRAVEL_Absolute);
+            }
+        }
+    }
+    if (Result != EOnJoinSessionCompleteResult::Success)
+    {
+        JoinButton->SetIsEnabled(true);
+    }
+}
+
+void UMenu::OnDestroySession(bool bWasSuccessful)
+{
+
+}
+
+void UMenu::OnStartSession(bool bWasSuccessful)
+{
+
+}
